+++ solutions/swb-reference/src/environment/ec2Rstudio/ec2RstudioLaunchSSM.yaml
@@ -67,6 +67,22 @@ parameters:
     type: String
     description: The IAM policy to be associated with the launched workstation
     allowedPattern: ^\{.*\}$
+  ALBSecurityGroup:
+    type: String
+    description: Security Group Id of the ALB to which the EC2 instance traffic will be allowed from
+    allowedPattern: ^sg-[a-zA-Z0-9]+$
+  ListenerArn:
+    type: String
+    description: ARN of the alb listener in which the rules will be created
+    allowedPattern: ^arn:aws:elasticloadbalancing:(us(-gov)?|ap|ca|cn|eu|sa)-(central|(north|south)?(east|west)?)-\d:\d{12}:listener\/[a-zA-Z0-9_\-\/]+$
+  ListenerRulePriority:
+    type: String
+    description: Priority of the rule to be created
+    allowedPattern: \d+
+  ApplicationUrl:
+    type: String
+    description: URL of the application. This will be used as the host header to forward to target group
+    allowedPattern: ^[a-zA-Z0-9_\-.]{1,100}$
   DatasetsBucketArn:
     type: String
     description: The ARN of the Datasets bucket in the main account
@@ -114,6 +130,14 @@ mainSteps:
           Value: '{{ VPC }}'
         - Key: Subnet
           Value: '{{ Subnet }}'
+        - Key: ALBSecurityGroup
+          Value: '{{ ALBSecurityGroup }}'
+        - Key: ListenerArn
+          Value: '{{ ListenerArn }}'
+        - Key: ListenerRulePriority
+          Value: '{{ ListenerRulePriority }}'
+        - Key: ApplicationUrl
+          Value: '{{ ApplicationUrl }}'
         - Key: EnvironmentInstanceFiles
           Value: '{{ EnvironmentInstanceFiles }}'
         - Key: EncryptionKeyArn
